Cleanup header files and include statements in fwdec server
Also formatted a little
