+++ minix/servers/fwdec/fwrule.c
@@ -1,11 +1,5 @@
 #include "fwrule.h"
 
-#include <stdint.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-
-#include "fwdec.h"
 #include "inc.h"
 
 /**
