+++ minix/servers/fwdec/main.c
@@ -1,9 +1,4 @@
 #include <minix/endpoint.h>
-#include <stdarg.h>
-#include <stdbool.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
 
 #include "inc.h" /* include master header file */
